projects/rag-expert-assistant/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
chromadb>=0.4.0
cohere>=5.0.0
sentence-transformers>=2.2.0
ragas>=0.1.0
datasets>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
openai>=1.0.0
pytest>=8.0.0